+++ src/Utils/PageToJson.go
@@ -8,13 +8,13 @@ import (
 )
 
 func WritePage(resultPath string, page *DataStructure.Page) {
-	outFile, err := os.Create(resultPath + page.PageID+".json")
+	outFile, err := os.Create(resultPath + page.PageID + ".json")
 	if err == nil {
 		writer := bufio.NewWriter(outFile)
 		defer outFile.Close()
 
 		var dictPage, err = json.Marshal(page)
-		if err == nil{
+		if err == nil {
 			_, _ = writer.Write(dictPage)
 			_ = writer.Flush()
 		}
@@ -22,27 +22,27 @@ func WritePage(resultPath string, page *DataStructure.Page) {
 }
 
 func WriteMappedPage(resultPath string, page *DataStructure.PageElement) {
-	outFile, err := os.Create(resultPath + "M"+page.PageId+".json")
+	outFile, err := os.Create(resultPath + "M" + page.PageId + ".json")
 	if err == nil {
 		writer := bufio.NewWriter(outFile)
 		defer outFile.Close()
 
 		var dictPage, err = json.Marshal(page)
-		if err == nil{
+		if err == nil {
 			_, _ = writer.Write(dictPage)
 			_ = writer.Flush()
 		}
 	}
 }
 
-func WriteGlobalWord(resultPath string, gloabalWord *map[string]int){
+func WriteGlobalWord(resultPath string, gloabalWord *map[string]uint64) {
 	outFile, err := os.Create(resultPath + "GlobalWord.json")
 	if err == nil {
 		writer := bufio.NewWriter(outFile)
 		defer outFile.Close()
 
 		var dictPage, err = json.Marshal(gloabalWord)
-		if err == nil{
+		if err == nil {
 			_, _ = writer.Write(dictPage)
 			_ = writer.Flush()
 		}
